backend/requirements.txt
adb_shell==0.4.4
alembic==1.11.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
argcomplete==2.1.2
attrs==25.3.0
Authlib==1.6.3
bcrypt==4.0.1
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
commonmark==0.9.1
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
distlib==0.4.0
dnspython==2.7.0
dparse==0.6.4
ecdsa==0.19.1
email-validator==2.0.0
et_xmlfile==2.0.0
fastapi==0.116.1
filelock==3.16.1
fonttools==4.59.1
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.14.0
httpcore==0.17.3
httptools==0.6.4
httpx==0.24.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
lxml==6.0.1
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
matplotlib==3.10.5
memory-profiler==0.61.0
narwhals==2.1.2
nltk==3.9.1
nox==2022.11.21
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
passlib==1.7.4
pillow==11.3.0
platformdirs==4.3.8
plotly==6.3.0
pluggy==1.6.0
protobuf==6.32.0
psutil==6.1.1
pure-python-adb==0.3.0.dev0
pyarrow==21.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pytest==7.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
render-cli==0.2.1
requests==2.32.5
rich==12.6.0
rpds-py==0.27.0
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
safety==3.6.0
safety-schemas==0.0.14
schedule==1.2.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
starlette==0.47.3
streamlit==1.48.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.22.0
virtualenv==20.34.0
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
